requirements.txt
tinytuya
requests
pyyaml
